this is a microservice

+++ app/Http/Controllers/TodoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Todo;
 use Carbon\Carbon;
+use App\Traits\ApiResponser;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -14,35 +15,74 @@ class TodoController extends Controller
      *
      * @return void
      */
-    public function __construct()
+    use ApiResponser;
+
+    public function __construct(Todo $todo)
     {
-        //
+        $this->todo = $todo;
     }
-    public function index() 
+    public function index($user) 
     {
-        $todos = Todo::all();
+        $todos = $this->todo::where("user_id", $user)->get();
         
-        return response()->json($todos);
+        return $this->successResponse($todos);
     }
-    public function store(Request $request)
+    public function store(Request $request, $user)
     {
-        $this->validate($request, [
+        
+        $rules = [
             'date' => 'required',
             'description' => 'required'
-        ]);
+        ];
         
-        $todo = Todo::create([
+        $this->validate($request, $rules);
+        
+        $todo = $this->todo::create([
+            'user_id' => $user,
             'date' => Carbon::parse($request->date)->format('Y-m-d'),
             'description' => $request->description,
         ]);
        
-        return response()->json($todo, Response::HTTP_OK);
+        return $this->successResponse($todo, Response::HTTP_CREATED);
     }
-    public function destory($id) 
+    public function show($user, $todo)
+    {   
+        $todo = $this->todo::where("id", $todo)
+                            ->where("user_id", $user)
+                            ->first();
+
+        return $this->successResponse($todo);
+       
+    }
+    public function update(Request $request, $user, $todo)
+    {   
+        $this->validate($request, [
+            'date' => 'required',
+            'description' => 'required',
+        ]);
+
+        $todo = $this->todo::where("id", $todo)
+                            ->where("user_id", $user)
+                            ->first();
+
+        if(is_null($todo)) {
+            return $this->errorResponse("No Todo Relates to this user!", 404);
+        }
+
+        $todo->update($request->all());
+       
+        return $this->successResponse($todo);
+    }
+    public function destory($user, $todo) 
     {
-        $todo = Todo::find($id);
+        $todo = $this->todo::where("id", $todo)
+                            ->where("user_id", $user)
+                            ->first();
+        if(is_null($todo)) {
+            return $this->errorResponse("No Todo Relates to this user!", 404);
+        }
         $todo->delete();
 
-        return response()->json("Successfully Deleted", Response::HTTP_OK);
+        return $this->successResponse($todo);
     }
 }
